mock file existance

+++ lib/shell/alias_checker.rb
@@ -3,6 +3,13 @@ class AliasChecker
 
   def initialize(filename)
     @filename = filename
+
+    if !File.exist?(File.expand_path(filename))
+      puts "No alias cache found."
+      puts "Better output will be generated if you run 'alias > ~/.aliases.cache'"
+      return
+    end
+
     file_contents = File.open(File.expand_path(filename), 'r').read
     lines = file_contents.split("\n")
     @aliases = {}
